chore: move buildApp to its own file
