+++ packages/deploy/src/deploy.ts
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-import chalk from 'chalk';
 import fs from 'fs-extra';
 import * as pulumi from '@pulumi/pulumi';
 import inquirer from 'inquirer';
@@ -24,12 +23,10 @@ import { AWSProgram } from './programs/aws';
 import { findPaths } from '@backstage/cli-common';
 import { Task } from './lib/tasks';
 import { basename, resolve } from 'path';
-import { exec as execCb } from 'child_process';
-import { promisify } from 'util';
+import { buildApp } from './lib/build';
 
 // eslint-disable-next-line no-restricted-syntax
 const paths = findPaths(__dirname);
-const exec = promisify(execCb);
 
 const createFile = async (fileName: string) => {
   const BASE_PATH_OF_DEV_FILES = 'templates/pulumi';
@@ -43,23 +40,6 @@ const createFile = async (fileName: string) => {
       throw new Error(`Failed to create file: ${error.message}`);
     });
   });
-};
-
-const buildApp = async () => {
-  process.chdir(paths.targetRoot);
-
-  const runCmd = async (cmd: string) => {
-    await Task.forItem('executing', cmd, async () => {
-      await exec(cmd).catch(error => {
-        process.stdout.write(error.stderr);
-        process.stdout.write(error.stdout);
-        throw new Error(`Could not execute command ${chalk.cyan(cmd)}`);
-      });
-    });
-  };
-
-  await runCmd('yarn tsc');
-  await runCmd('yarn build:backend');
 };
 
 export default async (opts: OptionValues) => {
